refactor: organize code by folders
